+++ wasp/tests/sv-comp/lib/assert.h
@@ -4,7 +4,7 @@
 void assume(int);
 void assert(int);
 int    sym_int    (char*);
-long   sym_long   (char*);
+long long  sym_long   (char*);
 float  sym_float  (char*);
 double sym_double (char*);
